move UI components to separate folder
