+++ frontend/src/components/ProjectForm.jsx
@@ -1,5 +1,5 @@
-import Button from "./Button";
-import Input from "./Input";
+import Button from "./UI/Button";
+import Input from "./UI/Input";
 
 const ProjectForm = ({ onSubmit, value, onChange, onClick, submitText }) => {
   return (
